strip to only s2n-tls and criterion bench
