+++ bindings/rust/bench/Cargo.toml
@@ -7,10 +7,6 @@ edition = "2021"
 
 [dependencies]
 s2n-tls = { version = "=0.0.32", path = "../s2n-tls", features = [ "testing" ] }
-rustls = "0.21.1"
-rustls-pemfile = "1.0.2"
-bytes = "1.4.0"
-openssl = { version = "0.10.54" }
 errno = "0.3.1"
 libc = "0.2"
 criterion = "0.3"
