+++ bindings/rust/bench/src/harness.rs
@@ -3,15 +3,6 @@ use std::{
     io::{Read, Write},
 };
 
-// TODO: figure out why 2 round trips are happening
-// TODO: use something other than Vec<u8> (BytesMut? VecDeque<Bytes>?)
-// TODO: change visibilities of functions
-// TODO: refactor to have common harness/add methods to custom harness
-// TODO: add comments
-// TODO: bulk transfer
-// TODO: understand cert generate script/customize
-// TODO: change rustls to tls1.3
-
 pub fn read_to_bytes(path: &str) -> Vec<u8> {
     read_to_string(path).unwrap().into_bytes()
 }
@@ -25,29 +16,11 @@ pub trait TlsImpl {
     /// Initialize buffers, configs, and connections (unhandshaked)
     fn new() -> Self;
 
-    /// Reinitialize connection with same configs
-    fn reinit(&mut self);
-
-    /// Process handshake for a single connection (i.e. client or server)
-    fn handshake_conn(&mut self, mode: Mode);
-
     /// Run handshake on initialized connection
-    fn handshake(&mut self) {
-        // set limit on round trips
-        let mut iter_remaining = 10;
-        while !self.has_handshaked() && iter_remaining > 0 {
-            self.handshake_conn(Mode::Client);
-            self.handshake_conn(Mode::Server);
-            iter_remaining -= 1;
-        }
-    }
+    fn handshake(&mut self);
 
     /// Checks if handshake is finished for both client and server
     fn has_handshaked(&self) -> bool;
-
-    /// Send and receive data from each connection.
-    /// If handshake hasn't happened yet, will handshake first
-    fn bulk_transfer(&mut self, data: &mut [u8]);
 }
 
 /// Read+Write buffer used for custom IO for different TlsImpl
