--- bindings/rust/rust-toolchain
@@ -1,1 +1,0 @@
-1.63.0
